requirements-optional.txt
# hard (runtime) dependencies
# keep this in sync with setup.py

# (only django, but the version is chosen by the tox environment)

# dependencies for tox testing
django-authtools
django-db-constraints
django-filter
django-storages
djangorestframework
logging_tree


# .. plus specific version dependencies defined in tox.ini
# we can't list them here because of https://github.com/pypa/pip/issues/2367
